Split HTML pages into separate headers
